requirements.txt
# T-Invest Python SDK (package index: https://opensource.tbank.ru/.../pypi/simple)
--extra-index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple
t-tech-investments>=0.3.4
